+++ cmd/kube-proxy/app/server.go
@@ -272,6 +272,7 @@ func NewProxyServerDefault(config *options.ProxyServerConfig) (*ProxyServer, err
 				loadBalancer,
 				net.ParseIP(config.BindAddress),
 				iptInterface,
+				execer,
 				*utilnet.ParsePortRangeOrDie(config.PortRange),
 				config.IPTablesSyncPeriod.Duration,
 				config.IPTablesMinSyncPeriod.Duration,
