+++ pkg/proxy/iptables/proxier.go
@@ -46,6 +46,7 @@ import (
 	"k8s.io/kubernetes/pkg/features"
 	"k8s.io/kubernetes/pkg/proxy"
 	"k8s.io/kubernetes/pkg/proxy/healthcheck"
+	utilproxy "k8s.io/kubernetes/pkg/proxy/util"
 	utilexec "k8s.io/kubernetes/pkg/util/exec"
 	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
 	utilsysctl "k8s.io/kubernetes/pkg/util/sysctl"
@@ -552,7 +553,7 @@ func (proxier *Proxier) OnServiceUpdate(allServices []api.Service) {
 		glog.V(4).Infof("Skipping proxy iptables rule sync on service update because nothing changed")
 	}
 
-	proxier.deleteServiceConnections(staleUDPServices.List())
+	utilproxy.DeleteServiceConnections(proxier.exec, staleUDPServices.List())
 }
 
 // Reconstruct the list of endpoint infos from the endpointIP list
@@ -792,7 +793,7 @@ func (proxier *Proxier) deleteEndpointConnections(connectionMap map[endpointServ
 		if svcInfo, ok := proxier.serviceMap[epSvcPair.servicePortName]; ok && svcInfo.protocol == api.ProtocolUDP {
 			endpointIP := strings.Split(epSvcPair.endpoint, ":")[0]
 			glog.V(2).Infof("Deleting connection tracking state for service IP %s, endpoint IP %s", svcInfo.clusterIP.String(), endpointIP)
-			err := proxier.execConntrackTool("-D", "--orig-dst", svcInfo.clusterIP.String(), "--dst-nat", endpointIP, "-p", "udp")
+			err := utilproxy.ExecConntrackTool(proxier.exec, "-D", "--orig-dst", svcInfo.clusterIP.String(), "--dst-nat", endpointIP, "-p", "udp")
 			if err != nil && !strings.Contains(err.Error(), noConnectionToDelete) {
 				// TODO: Better handling for deletion failure. When failure occur, stale udp connection may not get flushed.
 				// These stale udp connection will keep black hole traffic. Making this a best effort operation for now, since it
@@ -801,33 +802,6 @@ func (proxier *Proxier) deleteEndpointConnections(connectionMap map[endpointServ
 			}
 		}
 	}
-}
-
-// deleteServiceConnection use conntrack-tool to delete UDP connection specified by service ip
-func (proxier *Proxier) deleteServiceConnections(svcIPs []string) {
-	for _, ip := range svcIPs {
-		glog.V(2).Infof("Deleting connection tracking state for service IP %s", ip)
-		err := proxier.execConntrackTool("-D", "--orig-dst", ip, "-p", "udp")
-		if err != nil && !strings.Contains(err.Error(), noConnectionToDelete) {
-			// TODO: Better handling for deletion failure. When failure occur, stale udp connection may not get flushed.
-			// These stale udp connection will keep black hole traffic. Making this a best effort operation for now, since it
-			// is expensive to baby sit all udp connections to kubernetes services.
-			glog.Errorf("conntrack return with error: %v", err)
-		}
-	}
-}
-
-//execConntrackTool executes conntrack tool using given parameters
-func (proxier *Proxier) execConntrackTool(parameters ...string) error {
-	conntrackPath, err := proxier.exec.LookPath("conntrack")
-	if err != nil {
-		return fmt.Errorf("Error looking for path of conntrack: %v", err)
-	}
-	output, err := proxier.exec.Command(conntrackPath, parameters...).CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("Conntrack command returned: %q, error message: %s", string(output), err)
-	}
-	return nil
 }
 
 // This is where all of the iptables-save/restore calls happen.
@@ -1392,10 +1366,9 @@ func (proxier *Proxier) syncProxyRules() {
 // https://github.com/docker/docker/issues/8795
 // https://github.com/kubernetes/kubernetes/issues/31983
 func (proxier *Proxier) clearUdpConntrackForPort(port int) {
-	var err error = nil
 	glog.V(2).Infof("Deleting conntrack entries for udp connections")
 	if port > 0 {
-		err = proxier.execConntrackTool("-D", "-p", "udp", "--dport", strconv.Itoa(port))
+		err := utilproxy.ExecConntrackTool(proxier.exec, "-D", "-p", "udp", "--dport", strconv.Itoa(port))
 		if err != nil && !strings.Contains(err.Error(), noConnectionToDelete) {
 			glog.Errorf("conntrack return with error: %v", err)
 		}
