+++ test/test_owners.csv
@@ -768,6 +768,7 @@ k8s.io/kubernetes/pkg/proxy/config,ixdy,1,
 k8s.io/kubernetes/pkg/proxy/healthcheck,rrati,0,
 k8s.io/kubernetes/pkg/proxy/iptables,freehan,0,
 k8s.io/kubernetes/pkg/proxy/userspace,luxas,1,
+k8s.io/kubernetes/pkg/proxy/util,knobunc,0,
 k8s.io/kubernetes/pkg/proxy/winuserspace,jbhurat,0,
 k8s.io/kubernetes/pkg/quota,sttts,1,
 k8s.io/kubernetes/pkg/quota/evaluator/core,yifan-gu,1,
